pyproject.toml
[project]
name = "maria-wish-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.9.0",
    "pydantic>=2.10",
    "asyncpg>=0.30.0",
    "apscheduler>=3.10.4",
    "python-dotenv>=1.0.1",
    "google-genai>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]
